requirements.txt
llama-index==0.12.0
llama-index-core==0.12.0
llama-index-llms-openai==0.3.0
llama-index-embeddings-openai==0.3.0
openai==1.54.0
python-dotenv==1.0.0
pypdf==5.1.0
tavily-python==0.5.0
deepgram-sdk==5.3.0
elevenlabs==1.11.0
pydantic==2.11.5
fastapi==0.115.6
uvicorn==0.32.1
websockets==14.1
numpy==1.26.4
